# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Profesional",
        "description": "API REST de usuarios con autenticación JWT construida con Slim 4, PDO y MySQL. Endpoints públicos para login y health check, endpoints protegidos con Bearer token para el CRUD completo de usuarios.",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@miapi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api_electrofix.localhost",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar todos los usuarios",
                "description": "Devuelve un array con todos los usuarios registrados. La contraseña nunca se incluye en la respuesta.",
                "operationId": "b255eaafc8700b6669c570bdcc7fb124",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Crea un usuario con nombre, email y contraseña. La contraseña se almacena hasheada con bcrypt. Requiere token JWT.",
                "operationId": "cd7d427618e3dfee204e02a9deeb53ab",
                "requestBody": {
                    "description": "Datos del nuevo usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "María García",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com",
                                        "maxLength": 100
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "mipass123",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Usuario creado"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos incompletos o inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Datos incompletos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Endpoint público que crea un nuevo usuario administrador. La contraseña se almacena hasheada con bcrypt.",
                "operationId": "b47bf127e78e28fa8afd1e0de08c5ac7",
                "requestBody": {
                    "description": "Datos del nuevo usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nuevo Admin",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuevo@example.com",
                                        "maxLength": 100
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "mipass123",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Usuario registrado"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos incompletos o email ya registrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El email ya está registrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Usuario",
                "description": "Representación de un usuario del sistema. Se usa tanto para las peticiones (crear/actualizar) como para las respuestas (listar/obtener).",
                "required": [
                    "nombre",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "Identificador único del usuario (autoincremental)",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre completo del usuario",
                        "type": "string",
                        "example": "Juan Pérez",
                        "maxLength": 100
                    },
                    "email": {
                        "description": "Correo electrónico (único en el sistema)",
                        "type": "string",
                        "format": "email",
                        "example": "juan@correo.com",
                        "maxLength": 100
                    },
                    "password": {
                        "description": "Contraseña del usuario (solo se usa en creación y actualización; nunca se devuelve en respuestas GET)",
                        "type": "string",
                        "format": "password",
                        "writeOnly": true,
                        "example": "miClaveSegura123",
                        "maxLength": 255
                    },
                    "created_at": {
                        "description": "Fecha de registro del usuario (formato ISO 8601)",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-03-01T10:30:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenido desde POST /login. Incluirlo en el header Authorization como: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}